LinkIn scrap/job-matcher-app/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.5.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
selenium==4.12.0
beautifulsoup4==4.12.2